api/app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.8
pydantic-settings==2.6.1
filelock==3.17.0
flower==2.0.1
fsspec==2025.2.0
ftfy==6.3.1
h11==0.14.0
humanize==4.12.1
idna==3.10
Jinja2==3.1.5
kombu==5.4.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
pandas==2.2.3
peewee==3.17.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
fastapi-utils==0.7.0
email-validator==2.1.1
# yt-dlp==2025.2.19
watchdog
librosa
debugpy
psycopg2-binary==2.9.9
pydub
faster_whisper
gunicorn
httpx
sse_starlette
hf_xet
# pytube
pytubefix
langdetect
python-multipart
python-jose[cryptography]
passlib[bcrypt]
jwt
torchaudio